refactor: 17017 Move normalization logic for `StateFileData` to `readStateFileData`
