+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/state/snapshot/SignedStateFileUtils.java
@@ -16,14 +16,10 @@
 
 package com.swirlds.platform.state.snapshot;
 
-import java.util.Set;
-
 /**
  * Utility methods for dealing with signed states on disk.
  */
 public final class SignedStateFileUtils {
-
-    public static final String SIGNATURE_SET_FILE_NAME = "signatureSet.bin";
 
     public static final String HASH_INFO_FILE_NAME = "hashInfo.txt";
 
@@ -31,16 +27,6 @@ public final class SignedStateFileUtils {
      * The name of the file that contains the human-readable address book in the saved state
      */
     public static final String CURRENT_ADDRESS_BOOK_FILE_NAME = "currentAddressBook.txt";
-
-    /**
-     * The initial version of the signature set file
-     */
-    public static final int INIT_SIG_SET_FILE_VERSION = 1;
-
-    /**
-     * The supported versions of the signature set file
-     */
-    public static final Set<Integer> SUPPORTED_SIGSET_VERSIONS = Set.of(INIT_SIG_SET_FILE_VERSION);
 
     private SignedStateFileUtils() {}
 }
